refactor: 체스 기물 관련 클래스 패키지 분리
