+++ src/main/java/chess/ChessBoard.java
@@ -1,5 +1,7 @@
 package chess;
 
+import chess.piece.*;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
