+++ src/main/java/chess/piece/Empty.java
@@ -1,4 +1,6 @@
-package chess;
+package chess.piece;
+
+import chess.Side;
 
 public class Empty extends ChessPiece{
