+++ src/main/java/chess/piece/King.java
@@ -1,4 +1,6 @@
-package chess;
+package chess.piece;
+
+import chess.Side;
 
 public class King extends ChessPiece{
