+++ src/main/java/chess/piece/Rook.java
@@ -1,4 +1,6 @@
-package chess;
+package chess.piece;
+
+import chess.Side;
 
 public class Rook extends ChessPiece{
     public Rook(Side side) {
