+++ src/test/java/chess/ChessPieceTest.java
@@ -1,5 +1,6 @@
 package chess;
 
+import chess.piece.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
